FDG-9929 description can use markdown
